Rename serializer files
File names should not be the same as their folder
