+++ app/stac_api/serializers/collection.py
@@ -8,12 +8,12 @@ from stac_api.models import Collection
 from stac_api.models import CollectionAsset
 from stac_api.models import CollectionLink
 from stac_api.models import Provider
-from stac_api.serializers.serializers_utils import AssetsDictSerializer
-from stac_api.serializers.serializers_utils import HrefField
-from stac_api.serializers.serializers_utils import NonNullModelSerializer
-from stac_api.serializers.serializers_utils import UpsertModelSerializerMixin
-from stac_api.serializers.serializers_utils import get_relation_links
-from stac_api.serializers.serializers_utils import update_or_create_links
+from stac_api.serializers.utils import AssetsDictSerializer
+from stac_api.serializers.utils import HrefField
+from stac_api.serializers.utils import NonNullModelSerializer
+from stac_api.serializers.utils import UpsertModelSerializerMixin
+from stac_api.serializers.utils import get_relation_links
+from stac_api.serializers.utils import update_or_create_links
 from stac_api.utils import get_stac_version
 from stac_api.utils import is_api_version_1
 from stac_api.utils import isoformat
