+++ app/stac_api/serializers/general.py
