+++ app/stac_api/serializers/item.py
@@ -9,12 +9,12 @@ from rest_framework_gis import serializers as gis_serializers
 from stac_api.models import Asset
 from stac_api.models import Item
 from stac_api.models import ItemLink
-from stac_api.serializers.serializers_utils import AssetsDictSerializer
-from stac_api.serializers.serializers_utils import HrefField
-from stac_api.serializers.serializers_utils import NonNullModelSerializer
-from stac_api.serializers.serializers_utils import UpsertModelSerializerMixin
-from stac_api.serializers.serializers_utils import get_relation_links
-from stac_api.serializers.serializers_utils import update_or_create_links
+from stac_api.serializers.utils import AssetsDictSerializer
+from stac_api.serializers.utils import HrefField
+from stac_api.serializers.utils import NonNullModelSerializer
+from stac_api.serializers.utils import UpsertModelSerializerMixin
+from stac_api.serializers.utils import get_relation_links
+from stac_api.serializers.utils import update_or_create_links
 from stac_api.utils import get_stac_version
 from stac_api.utils import is_api_version_1
 from stac_api.validators import normalize_and_validate_media_type
